Making Track and Waypoint use the TripStatisticsData class. This grealy simplifies those classes since the fields are the same. There's certainly more cleanup to do on top of this, but I thought this was already more than enough for a single change.
The only *functional* change expected from this is that the average
speed is now calculated over the time until the last known position,
which IMHO makes sense since the distance used in that calculation only
includes up to that position. The visible effect is that having no GPS
signal doesn't make the speed decrease.

This was imported from perforce CL 15953263.